Byte -> Int

+++ src/main/java/domain/NobleCard.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class NobleCard extends Card {
     private final List<GemAmount> price;
 
-    public NobleCard(byte prestige) {
+    public NobleCard(int prestige) {
         super (prestige);
         this.price = new ArrayList<>();
     }
